pressley/requirements.txt
Django==1.4
MySQL-python==1.2.3
argparse==1.2.1
certifi==0.0.8
chardet==1.0.1
django-piston==0.2.3
feedparser==5.1.2
gevent==0.13.7
greenlet==0.4.0
httplib2==0.7.4
lxml==2.3.4
oauthlib==0.1.3
pyasn1==0.1.3
python-dateutil==2.1
git+git://github.com/sunlightlabs/python-superfastmatch.git@master#egg=python_superfastmatch
readability-lxml==0.2.5
requests==0.12.1
rsa==3.0.1
six==1.1.0
stream==0.8
uWSGI==1.2.4
wsgiref==0.1.2
pytz==2013.9
